mes-5057: consolidate test schema union types
