+++ src/modules/tests/test-data/cat-c/test-data.cat-c.selector.ts
@@ -1,40 +1,18 @@
 import { Competencies, LegalRequirements } from '../test-data.constants';
 import { get } from 'lodash';
 import { CompetencyOutcome } from '../../../../shared/models/competency-outcome';
-import { CatCUniqueTypes } from '@dvsa/mes-test-schema/categories/C';
-import { CatC1UniqueTypes } from '@dvsa/mes-test-schema/categories/C1';
-import { CatCEUniqueTypes } from '@dvsa/mes-test-schema/categories/CE';
-import { CatC1EUniqueTypes } from '@dvsa/mes-test-schema/categories/C1E';
+import {
+  CatCTestData,
+  CatCManoeuvres,
+  CatCTestRequirements,
+  CatCVehicleChecks,
+} from '../../../../shared/unions/test-schema-unions';
 import {
   NUMBER_OF_SHOW_ME_QUESTIONS,
 } from '../../../../shared/constants/show-me-questions/show-me-questions.vocational.constants';
 import {
    NUMBER_OF_TELL_ME_QUESTIONS,
   } from '../../../../shared/constants/tell-me-questions/tell-me-questions.vocational.constants';
-
-export type CatCTestData =
-  | CatCUniqueTypes.TestData
-  | CatC1UniqueTypes.TestData
-  | CatCEUniqueTypes.TestData
-  | CatC1EUniqueTypes.TestData;
-
-export type CatCManoeuvres =
-  | CatCUniqueTypes.Manoeuvres
-  | CatC1UniqueTypes.Manoeuvres
-  | CatCEUniqueTypes.Manoeuvres
-  | CatC1EUniqueTypes.Manoeuvres;
-
-export type CatCTestRequirements =
-  | CatCUniqueTypes.TestRequirements
-  | CatC1UniqueTypes.TestRequirements
-  | CatCEUniqueTypes.TestRequirements
-  | CatC1EUniqueTypes.TestRequirements;
-
-export type CatCVehicleChecks =
-  | CatCUniqueTypes.VehicleChecks
-  | CatC1UniqueTypes.VehicleChecks
-  | CatCEUniqueTypes.VehicleChecks
-  | CatC1EUniqueTypes.VehicleChecks;
 
 export const getDrivingFaultCount = (
   data: CatCTestData, competency: Competencies) => data.drivingFaults[competency];
