+++ src/modules/tests/test-data/cat-home-test/test-data.cat-home.selector.ts
@@ -1,7 +1,8 @@
-import { CatFUniqueTypes } from '@dvsa/mes-test-schema/categories/F';
-import { CatGUniqueTypes } from '@dvsa/mes-test-schema/categories/G';
-import { CatHUniqueTypes } from '@dvsa/mes-test-schema/categories/H';
-import { CatKUniqueTypes } from '@dvsa/mes-test-schema/categories/K';
+import {
+  CatHomeTestData,
+  CatHomeTestManoeuvres,
+  CatHomeTestVehicleChecks,
+} from '../../../../shared/unions/test-schema-unions';
 import {
   NUMBER_OF_SHOW_ME_QUESTIONS,
 } from '../../../../shared/constants/show-me-questions/show-me-questions.cat-home-test.constants';
@@ -9,23 +10,6 @@ import {
   NUMBER_OF_TELL_ME_QUESTIONS,
 } from '../../../../shared/constants/tell-me-questions/tell-me-questions.cat-home-test.constants';
 import { get } from 'lodash';
-
-export type CatHomeTestData =
-  | CatFUniqueTypes.TestData
-  | CatGUniqueTypes.TestData
-  | CatHUniqueTypes.TestData
-  | CatKUniqueTypes.TestData;
-
-export type CatHomeTestManoeuvres =
-| CatFUniqueTypes.Manoeuvres
-| CatGUniqueTypes.Manoeuvres
-| CatHUniqueTypes.Manoeuvres;
-
-export type CatHomeTestVehicleChecks =
-  | CatFUniqueTypes.VehicleChecks
-  | CatGUniqueTypes.VehicleChecks
-  | CatHUniqueTypes.VehicleChecks
-  | CatKUniqueTypes.VehicleChecks;
 
 export const getVehicleChecks = (
   state: CatHomeTestData): CatHomeTestVehicleChecks => state.vehicleChecks;
