+++ src/modules/tests/test-data/common/highway-code-safety/highway-code-safety.selectors.ts
@@ -1,4 +1,4 @@
-import { HighwayCodeSafetyUnion } from './highway-code-safety.reducer';
+import { HighwayCodeSafetyUnion } from '../../../../../shared/unions/test-schema-unions';
 import { get } from 'lodash';
 
 export const isHighwayCodeSafetySelected =
